Revert "network debug"
This reverts commit 0bacd980b5787a1dc2133822bbba9f92ccc431fd.

+++ src/core/network.h
@@ -53,7 +53,7 @@ private:
     QTimer *timerPing;
     qint64 iActive;
     QList<QString> msgSendQueue;
-    QHash<int, NetworkStats> msgSendHistory;
+    QHash<qint64, NetworkStats> msgSendHistory;
     QMutex mutex;
     bool bAuthorized;
